Rename new find-symmetry-lines
