+++ src/find-symmetry-lines.ts
