+++ src/index.ts
@@ -1,4 +1,4 @@
-import { findSymmetryLines } from './find-symmetry-lines-2';
+import { findSymmetryLines } from './find-symmetry-lines';
 import { strict as assert } from 'assert';
 import { Point } from './interfaces';
 import { rotatePoint } from './rotate-point';
